Move tile_dist and scoredist to spatial now that they're not used in coalesce anymore
